requirements.txt
torchvision
sacred
pytorch-utils
incense
pyrsistent
